no-issue: add controls plugin and refactor some of button components to use new syntax and args
