+++ packages/desktop/stories/buttons.stories.js
@@ -1,96 +1,39 @@
 import React from 'react';
-import styled from 'styled-components';
-import { storiesOf } from '@storybook/react';
-import { action } from '@storybook/addon-actions';
 import { Button, OutlinedButton } from '../app/components/Button';
-import { ButtonRow, ConfirmCancelRow } from '../app/components/ButtonRow';
-import { DropdownButton } from '../app/components/DropdownButton';
 
-storiesOf('Buttons/ButtonRow', module)
-  .add('Default', () => <Button variant="contained">Button</Button>)
-  .add('Primary', () => (
-    <Button variant="contained" color="primary">
-      Button
-    </Button>
-  ))
-  .add('Outlined', () => <OutlinedButton>Button</OutlinedButton>)
-  .add('ConfirmCancel', () => (
-    <ConfirmCancelRow onConfirm={action('confirm')} onCancel={action('confirm')} />
-  ))
-  .add('With custom text', () => (
-    <ConfirmCancelRow
-      onConfirm={action('confirm')}
-      onCancel={action('confirm')}
-      confirmText="OK"
-      cancelText="Back"
-    />
-  ))
-  .add('With long custom text', () => (
-    <ConfirmCancelRow
-      onConfirm={action('confirm')}
-      onCancel={action('confirm')}
-      confirmText="Assign patient diagnosis"
-      cancelText="Return to previous state"
-    />
-  ))
-  .add('With custom buttons', () => (
-    <ButtonRow>
-      <Button onClick={action('plier')} variant="contained" color="primary">
-        Plier
-      </Button>
-      <Button onClick={action('etendre')} variant="contained" color="secondary">
-        Etendre
-      </Button>
-      <Button onClick={action('relever')} variant="contained">
-        Relever
-      </Button>
-      <Button onClick={action('glisser')} variant="contained">
-        Glisser
-      </Button>
-    </ButtonRow>
-  ));
+export default {
+  argTypes: {
+    color: { control: 'select', options: ['primary', 'secondary'] },
+    variant: {
+      control: 'select',
+      options: ['contained', 'outlined', 'text'],
+    },
+    isSubmitting: {
+      control: 'boolean',
+    },
+    disabled: {
+      control: 'boolean',
+    },
+    hasPermission: {
+      control: 'boolean',
+    },
+  },
+  title: 'Button',
+  component: Button,
+};
 
-const actions = [
-  { label: 'button', onClick: () => {} },
-  { label: 'Etendre', onClick: () => {} },
-  { label: 'Relever', onClick: () => {} },
-  { label: 'Glisser', onClick: () => {} },
-];
+const Template = args => <Button {...args} />;
 
-const Container = styled.div`
-  display: flex;
-  align-items: flex-end;
-  justify-content: flex-start;
-  padding: 1rem;
+export const Primary = Template.bind({});
+Primary.args = {
+  color: 'primary',
+  variant: 'contained',
+  children: 'Button',
+};
 
-  > div {
-    margin-right: 18px;
-    margin-bottom: 3px;
-  }
-`;
-
-storiesOf('Buttons/DropdownButton', module)
-  .add('Default', () => (
-    <div>
-      <Container>
-        <DropdownButton actions={actions} size="large" />
-        <DropdownButton actions={actions} />
-        <DropdownButton actions={actions} size="small" />
-      </Container>
-      <Container>
-        <DropdownButton actions={actions} variant="outlined" size="large" />
-        <DropdownButton actions={actions} variant="outlined" />
-        <DropdownButton actions={actions} variant="outlined" size="small" />
-      </Container>
-    </div>
-  ))
-  .add('Only one action', () => (
-    <DropdownButton actions={[{ label: 'Plier', onClick: () => {} }]} />
-  ))
-  .add('No actions', () => <DropdownButton actions={[]} />)
-  .add('In button row', () => (
-    <ButtonRow>
-      <Button onClick={() => {}}>Other</Button>
-      <DropdownButton actions={actions} />
-    </ButtonRow>
-  ));
+const OutlinedTemplate = args => <OutlinedButton {...args} />;
+export const Outlined = OutlinedTemplate.bind({});
+Outlined.args = {
+  color: 'primary',
+  children: 'Button',
+};
